// dashboard/src/client/models/workflow-package-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { WorkflowPackageVersion } from './workflow-package-version';

export interface WorkflowPackageSummary {
    'id': string;
    'slug': string;
    'displayName': string;
    'summary'?: string | null;
    /**
     * Visibility policy (private/internal/public).
     */
    'visibility': string;
    'tags'?: Array<string> | null;
    'ownerId'?: string | null;
    'ownerName'?: string | null;
    'updatedAt'?: string | null;
    'latestVersion'?: WorkflowPackageVersion;
    /**
     * Base64-encoded preview from the latest version when available.
     */
    'previewImage'?: string | null;
}