// dashboard/src/client/models/workflow-publish-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface WorkflowPublishRequest {
    /**
     * Existing package id to append a version to.
     */
    'packageId'?: string;
    /**
     * Slug used when creating a new package.
     */
    'slug'?: string;
    'displayName'?: string;
    'summary'?: string | null;
    /**
     * Desired visibility (private/internal/public).
     */
    'visibility'?: string;
    'tags'?: Array<string> | null;
    /**
     * Semantic version for the published snapshot.
     */
    'version': string;
    'changelog'?: string | null;
    /**
     * Base64-encoded preview captured from the workflow canvas.
     */
    'previewImage'?: string | null;
}